requirements.txt
python-dotenv
langchain
langchain-community
pypdf
faiss-cpu
streamlit
llama-cpp-python
sentence-transformers
